Use createChooser on Android API level 18
Android API level 18 doesn't have a combined document picker, the intent opens apps directly. When the user chooses an app to use the intent with and uses the "Always" option they will not be able to pick a document from a different application in the future.

A chooser intent forces the choose window to always let you choose the app to pick documents from instead of going directly to one app all the time.