Replace UIDocumentMenuViewController with UIDocumentPickerViewController
UIDocumentMenuViewController is now deprecated because iOS 11 has a new interface for picking files from old providers and a new file provider api which does not need a provider to be picked before opening it.

Before iOS 11 directly opening a UIDocumentPickerViewController instead of going through a UIDocumentMenuViewController would just open the most recently opened document provider. This was considered an optional alternative method of launching a document picker because the user is still able to switch between document providers even without the document provider menu.

Because the menu is no longer used the iPad specific top/left properties are no longer relevant.