[UWP] Update Newtonsoft.Json dependency
Update project.json to point to React's version of Newtonsoft.Json

After upgrading react-native-windows to 0.57.0-rc.0 there was a conflict upstream with React itself having a newer version of this library. To remedy it I bumped the version number in the project.json. Then closed and reopened (likely restoring Nuget packages) and it fixed itself. Without this patch it refuses to compile.

That said there is some alternate way to get rid of this project.json entirely but I'm unaware of how to do this properly. So this is my simple patch in the meantime.